refactor appRevison extract func to utils
